DEBUG and Version REST API resources. Data to be generated automatically upon app launch in case no data in Flights table.
