+++ src/main/java/com/aerotravel/flightticketbooking/rest/v0/RestExceptionHandler.java
@@ -6,12 +6,16 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageConversionException;
+import org.springframework.validation.BindException;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 import javax.validation.ConstraintViolationException;
+import java.util.HashMap;
 
 @Slf4j
 @ControllerAdvice
@@ -26,10 +30,29 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
     }
 
     @ExceptionHandler({ConstraintViolationException.class,
-            DataIntegrityViolationException.class})
+            DataIntegrityViolationException.class, HttpMessageConversionException.class})
     public ResponseEntity<Object> handleBadRequest(
             Exception ex, WebRequest request) {
-        return handleExceptionInternal(ex, ex.getLocalizedMessage(),
+
+        var body = ex.getLocalizedMessage() + "\n" + ex.getMessage();
+
+        return handleExceptionInternal(ex, body,
                 new HttpHeaders(), HttpStatus.BAD_REQUEST, request);
+    }
+
+    @Override
+    protected ResponseEntity<Object> handleBindException(
+            BindException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
+
+        var errors = new HashMap<String, String>();
+        ex.getBindingResult().getAllErrors().forEach((error) -> {
+            var fieldName = ((FieldError) error).getField();
+            var errorMessage = error.getDefaultMessage();
+            errors.put(fieldName, errorMessage);
+        });
+
+        var body = ex.getLocalizedMessage() + "\n" + errors;
+
+        return handleExceptionInternal(ex, body, headers, status, request);
     }
 }
